test: add TransaoRepository tests

+++ src/main/java/dev/thiagorodrigues/carteira/application/dtos/ItemCarteiraProjection.java
@@ -1,11 +1,13 @@
 package dev.thiagorodrigues.carteira.application.dtos;
 
+import java.math.BigDecimal;
+
 public interface ItemCarteiraProjection {
 
     String getTicker();
 
     Long getQuantidade();
 
-    Double getPercentual();
+    BigDecimal getPercentual();
 
 }
